refactor: upd. types
