+++ screens/SendModalScreens/type.ts
@@ -1,0 +1,8 @@
+export type IPerson = {
+  _id: string;
+  avatar: any;
+  firstName: string;
+  lastName: string;
+};
+
+export type IMessage = {};
